[Browser Lib] Updated versions of node modules
:Release Notes:
Updated versions of node modules

:Detailed Notes:
Versions of node modules which have security vulnerabilities are found in
package-lock.json file and those modules are not actually used now.
This commit updates package info to avoid security alarms due to unused modules.
Also, this commit updates other modules to up-to-date versions which are
compatible with existing versions.
Any changes by this commit does not affect the behavior of the app,
so nothing should be changed if the app is built and launched well.

:Testing Performed:
Build and Run the browser

:QA Notes:

:Issues Addressed:
[PLAT-80793] [Enact browser] Update package-lock.json to avoid security alert

Change-Id: I5160c7765607d588a0d248ac681cf8d4433f83a8
Reviewed-on: http://gpro.lge.com/245433
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Yuri Kim <yurik.kim@lge.com>
Reviewed-by: Seungcheon Baek <sc.baek@lge.com>
Tested-by: Seungcheon Baek <sc.baek@lge.com>