[Enact Browser] Fix white screen at the first launch after flashing
:Release Notes:
Fixed abnormal white screen at the first launching

:Detailed Notes:
After site filtering feature is impelmented,
when the browser app is launched for the first time from a device,
the screen becomes white and no longer works right after typing
any letter in omnibox due to uninitialized state.
This fix is to add a guard code to handle uninitialized state.

:Testing Performed:
Build and Run the browser

:QA Notes:

:Issues Addressed:
[PLAT-80895] [Enact browser] the screen became white when type any letter in omnibox on the first launching

Change-Id: I4d6b3bf988bbd8622332285a3ee965541f428a0e
Reviewed-on: http://gpro.lge.com/245905
Tested-by: Seungcheon Baek <sc.baek@lge.com>
Reviewed-by: Gyeorye Lee <gyeorye.lee@lge.com>
Reviewed-by: Seungcheon Baek <sc.baek@lge.com>