[Enact browser] Fix omnibox to get focus to suggestions by tapped once
:Release Notes:
Fix omnibox to move focus to suggestions when a user taps one of
suggestions once.

:Detailed Notes:
We added new config to the component in order to solve focus issue which
a user taps other component twice to move focus while input field has
focus. (see https://github.com/enactjs/enact/pull/2660)
In this patch, we apply the config, {unlockedPointer: true}, to
HOC of InputSpotlightDecorator. For this, we also replace a customized
Input component insteadof moonstone/Input component for Omnibox.

:Testing Performed:
1. Launch web browser
2. Open any site started with 'g'. e.g, https://www.google.com
3. Open new tab
4. Enter 'g' in Omnibox
5. Tap any suggestion in the list.
6. Navigate the selected site

:QA Notes:

:Issues Addressed:
[PLAT-87249] No response when selecting address list

Change-Id: I3a6a3a11fd286a8da359196bcbe9e4d6a53bff6e