[Enact browser] Add a favorite URL to the launcher
:Release Notes:
Added a favorite URL to the launcher

:Detailed Notes:
When a user added a favorite URL from the web browser,
a dialog will show up and ask if he/she wants to add
favorite to the home.
If yes, add a new launch point to the home with the title
and the url of the website.
If not, add favorite only to web browser.

:Testing Performed:
1. Launch web browser
2. Go to website to add favorite(e.g. www.google.com)
3. Click on the 'add to bookmark' icon and see if a dialog
pops up
4. Click on 'yes' to add favorite to home
5. Go to home
6. See if the launch point was added with title

:QA Notes:

:Issues Addressed:
[PLAT-98920] Add a favorite URL to the launcher

Change-Id: Ida2ba6aad25641df7c3e28722846acd7496607e5