[Enact Browser] Open an url in a new tab when webOSRelaunch event detected
:Release Notes:
Open an url in a new tab when webOSRelaunch event detected

:Detailed Notes:
The custom event, webOSRelaunch, is created and triggered from chromium
app-shell. Enact browser opens a given url in the new tab or in the last
tab when webOSRelaunch event is triggered with url as custom data.

:Testing Performed:
QWA-2624 on webOS Auto
Test on PC
1. Run Enact browser
2. Open inspector
3. Execute the folloing codes on the console
  var event = new CustomEvent("webOSRelaunch", {detail: {url: "google.com"}});
  document.dispatchEvent(event);
4.Open google page in a new tab.

:QA Notes:
If the maxium tabs(7) are opened, the url of the last tab is changed
instread of opening new tab when relaunch event detected.

:Issues Addressed:
[PLAT-101402] [webOS Auto] Enact browser: Doesn't relaunch with "target" url
              in launch param

Change-Id: Iafb45f24eaf96cc7c361c682552b5d636eedecba