[Enact Browser] Fix an error page blinking when it reloaded
:Release Notes:
Fix an error page blinking when it reloaded

:Detailed Notes:
The enact browser provides the following two different kinds of error
pages; 1) An error page from chromium called built-in error page,
2) an error page from the application UI. we can choose which error page
is used by the 'useBuiltInErrorPages' property.
Although browser set the useBuiltInErrorPages as false in the application,
the builtin error page is shown shortly while the page is reloaded.
In this fatch, we fixed it.

:Testing Performed:
1. Launch the web Browser
2. Open the invalid page ex. https://www.errrrr.org/
3. An error page as application UI is shown.
4. Tab 'reload' button in the omnibox

:QA Notes:

:Issues Addressed:
[PLAT-106421] Blinking appears when reloads the invalid URL

Change-Id: Iae559ae4532b7a67d04d2430683b76f1b854b366