[Enact Browser] Fix the end of pages gets cut when bookmarkbar is shown
:Release Notes:
Fix the end of pages gets cut when bookmarkbar is shown

:Detailed Notes:
Some of pages like settings, bookmark manager and history looks cutting
the botton of the page.
In this patch, we adjust the height of scroller or list in the each page
to fit the their contents in the screen regardless whether bookmarkbar
is shown or not.

:Testing Performed:
1. Launch the web Browser
2. Tap menu button > 'Settings'
3. Turn on 'Always show Bookmars Bar'
4. Check a scroller of bookmark manager page and history page either

:QA Notes:

:Issues Addressed:
[PLAT-105621] "Clear Browsing Data" button gets cut when turn on
"Always show Bookmarks Bar" button in web Browser setting page.

Change-Id: I2b1ef9e3a8bec6046d6bb3f8f53659bada5b7080