[Enact Browser] Fix a suggestion popup to be invisible when no search word
:Release Notes:
Fix a suggestion popup to be invisible when no search word with while Omnibox
has focus

:Detailed Notes:
In this patch, we fix an omnibox not to lose focus if there is no search
word but still has focus on it.
This is a workaround code because ContextualPopupDecorator forces to shift
focus to the container node when popup is closed. we temporally hide the
suggestion popup to solve it when input value is changed.

:Testing Performed:
1. Open web browser
2. Type some letters on the Omnibox
3. Erase all letters
4. Cursor keeps alive in the Omnibox but the suggestion popup disappears

:QA Notes:

:Issues Addressed:
[PLAT-105607] VKB closes and fails to open after a word is selected from
Word Suggestion

Change-Id: I0c8cd3dfb775cd981f0a06566cce40df00b8bf8d
(cherry picked from commit 8339dbf530f2c03878d26885b05d7a7123662f35)