[Browser Lib] Add new method to lose the focus of text input in webview
:Release Notes:
Add new method to lose the focus if active element of document loaded
to webview is any kind of input.

:Detailed Notes:
When changing tab, some elements for inputting text has focus, it should
lose its focus for hiding virtual keyboard. So, in this patch, we add a
new method in webview mixin class to lose focus if the active element in
document of webview is a kind of text input.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-107958] VKB appeared on browser settings page when input field is not focused

Change-Id: I1288e38514aa64739d242729964fcbbbb86788f8