Revert "[webos][build][browser] Add OSS package information file"
:Release Notes:
Reverted the adding of OSS package information file.
Will be applied another solution on meta layer only.

:Detailed Notes:
This reverts commit 4b1682ae0f517c63b5e165f5c1fae87d271f4cb5.
Previous solution was based on adding OSS (Open Source Software) file into the
browser sources and updating Neva meta with proper checksum of that file.
But actually we don't need to have OSS file on Neva side.
And it is much simpler and more proper to just remove the check for OSS file if
it present in the com.webos.app.enactbrowser.bb file (like in OSE build and
thud branch of build-starfish).
Hence OSS file is remove in favor of the change on meta layer only
http://repo.lge.com:8080/#/c/neva/meta-lg-webos-neva/+/3990/.

:Testing Performed:
N/A

:QA Notes:

:Issues Addressed:
[NEVA-5234] Prepare infrastructure for webOS/OSE build with Neva

Change-Id: Ia4eace5d3b44fbc8b40da3444d24937a8322c238
Reviewed-on: http://repo.lge.com:8080/4034
Reviewed-by: <ivan.murashov@lge.com>
Release-Notes-Ok: <ivan.murashov@lge.com>
Tags-Ok: <ivan.murashov@lge.com>
Tested-by: <ivan.murashov@lge.com>