[browser][appshell] Hide the suggested URL list when clicking on webview content area
:Release Notes:
Hide the suggested URL list when clicking on webview content area

:Detailed Notes:
When users click on the input address bar, the suggested URL list
appears.
However, it is not hidden when users click outside of the address bar input.
This is because the browser div wrapper only listens for click events
on itself and its children.
The iframe in the webview does not belong to the aforementioned div,
so it cannot receive click events when users click on the webview content.
To fix this, we can use the 'window.postMessage' method to send the
click event from the webview iframe to its parent

:Testing Performed:
1. Start browser:
luna-send -n 1 -f luna://com.webos.applicationManager/launch '{"id": "com.webos.app.enactbrowser", "params":{"displayAffinity": 0,"target":"https://www.webex.com"}}'
2. Click to the input address bar and enter any characters
3. After the suggested URL list appeared, move the mouse to the
"webex.com" content page then click to an empty space on it.
4. The suggested URL list has disappeared

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRP-15775] [OSE RPi4 - VCS][Webex] The list does not disappear when you touch an empty space after entering another address in the address bar during Meeting

Change-Id: I92e8c40be80418faba396a7fbc133a3af0d5b090