[browser][appshell] Fix spotlight freezing when the cursor moves outside the address bar
:Release Notes:
Fixed spotlight freezing when the cursor moves outside the address bar
to restore spotlight moving backward/forward

:Detailed Notes:
After the website is loaded, click to the address bar, then use the
PKB to move the cursor to the last position of the URL.
Continue pressing the Right key once, and the spotlight will appear on
the bookmark icon.
However, we cannot use the navigation keys to change the spotlight to
other elements.
The root cause is that the spotlight is terminated whenever a new
webview tab is opened, and it cannot be resumed if the current element
is an input.
The solution is to enable spotlight before the cursor moves outside the
address bar

:Testing Performed:
1. Start browser:
luna-send -n 1 -f luna://com.webos.applicationManager/launch '{"id": "com.webos.app.enactbrowser", "params":{"displayAffinity": 0,"target":"https://www.google.com"}}'
2. Click to the input address bar and the VKB will appear
3. (Using the physical keyboard) Press the Right key once, then press
the Up key multiple times until the VKB disappears.
4. (Using the physical keyboard) Press the Right key until the cursor
moves to the last position of the URL.
5. (Using the physical keyboard) Continue pressing the Right key
multiple times, then press the Left key multiple times.
Obeservation: The spotlight is switched between elements.

:QA Notes:
Test-cases: [QWTC-12887] Spotlight working properly when use physical keyboard

:Issues Addressed:
[WRP-19487] [OSE-RPi4]: The cursor is not retrieving backward once cursor is on "bookmark" icon

Change-Id: Icb9f99249ebd8b0eba09f7244113b4734f110b8c