[browser][appshell] Added "Reset Settings" feature
:Release Notes:
Allow the user to restore the settings to default

:Detailed Notes:
This patch is a cherry-pick from the original patch
http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/372725
and updated the logic to suit the browser-shell.

Add Reset Settings button to the Settings page.
When the user clicks it, the popup confirmation shows to confirm Yes
or No for reseting.
Continue clicks Yes button, all the settings will be restored to
default value
 - startupPage: newTabPage
 - homePageUrl: www.google.com
 - searchEngine: Google
 - alwaysShowBookmarks: false
 - siteFiltering: off (Approved sites and blocked sites will be cleared)
 - privateBrowsing: false (close all current web tabs if the browser state is private mode)
 - pinNumber: 0000
All cookies data will be cleared.
All site filtering URLs will be cleared.

This patch requires a dependency from webruntime:
http://gpro.lge.com/c/webosose/chromium108/+/375846

:Testing Performed:
1. Launch Web Browser, then visit to Settings page
2. Change some settings:
 - On startup: Home page and edit home page to any sites
 - Search Engines: Bing
 - Always show Bookmarks Bar: On
 - Private Browsing: On
 - Site Filtering: Approved mode and add any sites
3. Click Reset Settings button
4. Click Yes after the confirmation dialog appears
5. Wait until the reseting is completed
Expected: All settings are restored to the default

:QA Notes:
Test case: QWTC-13246

:Issues Addressed:
[WRP-26941] Implement the code and logic changes in the ReactJS site
[WRQ-8483] Cherry-pick all patches of the reset setings to default feature from ReactJS

Change-Id: I9f273cb31692bc70ab397246ba7df58d55176b99