[browser][appshell] Hide the zoom, menu dropdown list
:Release Notes:
Hide the zoom, menu dropdown list  when clicking on webview content area

:Detailed Notes:
When users click on the zoom icon or menu icon, the dropdown list
corresponding appears.
However, they are not hidden when users click to webview content area.
This is because the browser div wrapper only listens for click events
on itself and its children.
The iframe in the webview does not belong to the aforementioned div,
so it cannot receive click events when users click on the webview content.
To fix this, we can use the 'window.postMessage' method to send the
click event from the webview iframe to its parent

:Testing Performed:
1. Start browser:
luna-send -n 1 -f luna://com.webos.applicationManager/launch '{"id": "com.webos.app.enactbrowser", "params":{"displayAffinity": 0,"target":"https://www.w3.org"}}'
2. Click to the zoom icon
3. After the zoom dropdown list appeared, move the mouse to the "w3.org"
content page then click to an empty space on it
4. Expect the zoom dropdown list has disappeared.
5. Click to the menu icon
6. After the menu dropdown list appeared, move the mouse to the "w3.org"
content page then click to an empty space on it
7. Expect the menu dropdown list has disappeared.

:QA Notes:
Test-cases: QWTC-13631

:Issues Addressed:
[WRQ-946] [browser][appshell] Disappear Zoom controller, Menu dropdown list when clicking to the webview content

Change-Id: I84dba62b8b808d116947905de541445fb83b4710