[browser][appshell] Fix the input address losing focus
:Release Notes:
Hold the VKB is always shown when the input address is being focused

:Detailed Notes:
The VKB is hidden while the input address is being focused, and this
occurs while the webpage is loading. The root cause of this issue is
that the webview element is being focused incorrectly due to incorrect
logic.

:Testing Performed:
1. Visit https://www.w3schools.com/jsref/tryit.asp?filename=tryjsref_confirm2
2. Click to the input address bar while the site is loading
3. Expect the vkb to always be shown even when the site has finished
loading.

:QA Notes:
Test-cases: QWTC-11115

:Issues Addressed:
[WRQ-3659] [TAS Failed][OSE][QWTC-11115] Verify dialog boxes work when one of alert, confirm, and prompt is called
[WRQ-4165] [TAS Failed][OSE Emulator][QWTC-11115] Verify dialog boxes work when one of alert, confirm, and prompt is called

Change-Id: I40573558546a7cf691b5aab43241a3c09fe1ad8d