[browser][appshell] Fix error can't open a new window from window.open event
:Release Notes:
The open new window event is not working because the creation of the
webview is blocked.

:Detailed Notes:
In the retrieving the custom user agent logic, the URL parameter may potentially be an invalid URL. This can lead to an error when using the `new URL(url)` in getUserAgentOverride method. To resolve this issue, a
try-catch block has been added to handle any potential errors.

:Testing Performed:
1. Visit http://neva.lge.com/test-resources/app/browser/popup/parent.html
2. Click "Open or send to Child_1 window" button
Expect a new webview tab is opened with url: http://neva.lge.com/test-resources/app/browser/popup/child1.html

:QA Notes:
Test-cases: QWTC-12288

:Issues Addressed:
[WRQ-4174] [TAS Failed][OSE Emulator][QWTC-12288] Verify tab-under feature for OAuth operation

Change-Id: I4f4eb8cebe950a5060a3566e4da1de25c500a5df