[browser][appshell] Fix bug the webview is not focusing properly
:Release Notes:
Focused to the webview instead of the back button when opening a site
by the luna command

:Detailed Notes:
When opening a site using luna command, the Back button is always
focused, and the tooltip overlaps the first tab.
This patchset make the webview is focused instead of the back button
in the first run app.

:Testing Performed:
1. Launch Enact browser
2. Open site: luna-send -n 1 -w 300000 luna://com.webos.service.applicationmanager/launch '{"id": "com.webos.app.en
actbrowser", "params": {"target": "https://www.google.com"}}'
3. Observer the back button is not focused and google site is opened
properly

:QA Notes:
Test-cases: QWTC-12221

:Issues Addressed:
[WRQ-5260] [TAS Failed][OSE][QWTC-12221] Verify default Zoom valueafter relaunching Browser app

Change-Id: I393bf19156ce0a190d1849a4f8ebe2cbd5ba8782