[browser][browsershell] Added User permission popup UI
:Release Notes:
Add Popup UI & Allow/Block buttons - Media, Geolocation permission popup

:Detailed Notes:
(1) UI for the popup that asks for the permission to access geolocation,
 camera/mic or both
(2) Allow,Block and close buttons for the user interaction
(3) API integration for media detect, allow/block/close functionality

This path is implemented based on the corresponding feature in the
app-shell. It is not a cherry-pick because the logic to show pop-up in
the browser-shell is based on PageView and IPC, instead of JS like in
the app-shell.

:Testing Performed:
1. Launch Web Browser, then visit https://mictests.com
2. Click "Test my mic" button
3. Permission popup appears
4. Click Allow/Block from popup
5. Permission popup disappears

:QA Notes:
Test case: QWTC-11895, QWTC-11893, QWTC-11740, QWTC-10174, QWTC-10173

Depends on: http://gpro.lge.com/c/webos-pro/chromium/+/391530

:Issues Addressed:
[WRQ-8492] Port the patches to be related to user permission pop-up
           UI (ReactJS)
[WRP-29343] [OSE-RPI4]:The User permission pop-up is not showing at
            the left corner of the browser
[WRP-14699] Implement userpermission pop-up UI for Geolocation
[WRO-13698] Popup UI & Allow/Block buttons implementation -
            EnactBrowser

Change-Id: I5804c689d8381351f383485c5998d098afa24ebb