[browser][browsershell] Fix the bookmark dialog does not appears
:Release Notes:
Fix the bookmark dialog does not appears when clicking the bookmark
icon to add/remove any site

:Detailed Notes:
Fix the bookmark dialog does not appears when clicking the bookmark
icon to add/remove any site because the feature is using an old method
to show/hide the UIOverlay. This patch updates the new method to
show/hide the UIOverlay

:Testing Performed:
1. Launch EnactBrowser
2. Visit https://yahoo.com
3. Click to Star icon, a successful bookmark dialog appears
4. After Yahoo site is bookmarked, click to Star icon again,
a successful 'remove bookmark' dialog appears.

:QA Notes:
Test cases: QWTC-12138

:Issues Addressed:
[WRQ-24244] The bookmark dialog does not appears when click
            add/remove bookmark icon

Change-Id: Id37521e723c7009dda3e5689956d759b786f3d30