[n_upsable][browser][browsershell] Fix the "Back" button does not work properly in the Pin code dialog
:Release Notes:
Fix the "Back" button does not work properly in the Pin code dialog

:Detailed Notes:
When the "Back" button is clicked, it triggers the onNumberKeyPressed
event, which handles two cases: "Back" or a number. The issue was
caused by checking for the wrong condition (BACK instead of "Back")
as the input key value.

:Testing Performed:
1. Start Enact Browser
2. Go to menu -> Settings
3. Click the Sitefilter button to display the Pin code dialog.
4. Enter the pin code and test the action of the "Back" button.

:QA Notes:
Test-cases: QWTC-14580

:Issues Addressed:
[WRQ-25347] The "Back" button does not work properly in the Pin
            code dialog.

Change-Id: I7cf96f5a9a6edda675ae79d0138759b4e5eedbfc