[n_upsable][browser][browsershell] Unnecessary update of the URL in the did-start-loading event
:Release Notes:
Unnecessary update of the URL of the navState in the did-start-loading
event

:Detailed Notes:
Unnecessary to update "url" state of the navState when the
did-start-loading event occurs. Instead, it should just update the
"isLoading" state.
Since "url" value (e.g. http://google.com) when the did-start-loading
event occurs is different from the value (e.g. https://www.google.com/)
when the did-finish-navigation event occurs, this lead to the history
store saving duplicate items.

:Testing Performed:
1. Launch Enact borwser
2. Open Settings page and Clear Browsing Data
3. Input url: google.com and press Enter to visit
4. Open new tab then enter url: wikipedia.org
4. Open History page
Observe that there are two history items in the list

:QA Notes:
Test cases: QWTC-13146

:Issues Addressed:
[WRQ-26013] [TAS Failed][OSE][QWTC-13146] Verify that visited web pages
            are displayed in chronological order.

Change-Id: I470427667dca13624aba76af1a4a8962ddfd05dd