[n_upsable][browser][browsershell] Always display the "X" button in the "Most Visited Sites".
:Release Notes:
Always display the "X" button in the "Most Visited Sites".

:Detailed Notes:
In the current version, the close button in the "Most Visited Sites" item only appears when hovering over the site item, which prevents
automation testing with test case QWTC-13145. So, it is always visible,
and add a black background to the button to ensure it remains visible
when the site's background is white.

:Testing Performed:
1. Start Enact Browser
2. Go to any page
3. Open a new tab
4. Remove a site from the most visited site

:QA Notes:
Test-cases: QWTC-13145

:Issues Addressed:
[WRQ-26012] [TAS Failed][OSE][QWTC-13145] Verify that remove a site
            from the most visited site.

Change-Id: I6d2b43e358318046f88b0cc0a47392575d51712d