[n_upsable][browser][browsershell] Fix error the pdf url is not shown in omnibox
:Release Notes:
Fix error the pdf url is not shown in omnibox

:Detailed Notes:
When opening a PDF through a Luna command, the PDF URL is not shown
in the omnibox. The "url" state of navState only updates with URLs
that start with /http|https|file/ when the did-finish-navigation
event occurs. Since the PDF URL loaded through the Google
extension starts with /chrome-extension/, it is not updated.

:Testing Performed:
1. Launch Enact Browser app
2. Launch pdf file through luna command
Observe that the omnibox properly displays the PDF URL

:QA Notes:
Test-cases: QWTC-12237

:Issues Addressed:
[TAS Failed][OSE][QWTC-12237] Verify that move to the bookmarks
manager when select the Bookmark Star icon in Bookmark bar

Change-Id: Ia96730b9e131ed2e0493ce7e7897ca5a33bdffa8