[n_upsable][browser][browsershell] Fix the urlSuggestions are not shown in the correct order
:Release Notes:
Fix error the urlSuggestions are not shown in any order

:Detailed Notes:
When accessing a site, the did-start-loading is called multiple
times. causing the hitCount(visit count) to increased inaccurately.
The urlSuggestions queried via hitCount return incorrect value.
To address this, the values will be saved when the main frame occurs the
did-finish-loading event. This ensures the db is updated only
once.

:Testing Performed:
1. Launch Enact Browser app
2. Open page: https://www.bing.com
3. Open page: https://www.naver.com
4. Open page: https://www.google.com
5. Open page: https://www.facebook.com
6. Open page: https://www.w3.org
7. Open new tab
8. Enter h in search bar
Observe the Input Suggestion List to ensure the URLs are displayd in
the order: w3, facebook, google, naver, bing

:QA Notes:
Test-cases: QWTC-13109

:Issues Addressed:
[WRQ-28483] [TAS Failed][OSE][QWTC-13109] Verify show list of 6 URL
            suggestions while entering URL.

Change-Id: I9c9f83c013ce56503e337e067d87b9469b7cf576
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/424019
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>