[n_upsable][browser][browsershell] Improve the behavior of the Menu, extensions, zoom button
:Release Notes:
Improve the behavior of the Menu, extensions, zoom button

:Detailed Notes:
The Menu dropdown list doesn't display properly when switching between
the Zoom button and the Menu button. When clicking on buttons that use
stopPropagation, the document click event is not triggered, and the
isOpened state is not updated. As a result, need to click twice to
change this state. The same issue occurs with other buttons such as
extensions and zoom button

:Testing Performed:
1. Launch the Enact Browser
2. Visit google.com
3. Click to Menu button to show the Menu dropdown list
4. Click to Zoom button to show the Zoom dropdown list
5. Click to Menu button once again
Observe the Menu dropdown list appears

:QA Notes:
Test-cases: QWTC-14801

:Issues Addressed:
[WRQ-32969] Improve the behavior of the Menu, extensions management
            and zoom button

Change-Id: If846187d18ef32397300e6f82a0fb98c57c7831a
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/428768
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>