[n_upsable][browser][browsershell] Pin numbers are not working properly when clicking near the center button
:Release Notes:
Pin numbers are not working properly when clicking near the center
button

:Detailed Notes:
Pin number buttons have default left and right padding. When clicking
on the padded area, the system doesn't receive the value from
ev.target.textContent of the pin number.
The border radius was removed because it was unnecessary.

:Testing Performed:
1. Launch the Enact Browser
2. Open Settings page
3. Click to Site Filtering toggle to show the Pin dialog
4. Click around the center digit buttons
Observe the digit appears in the textbox

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRQ-33083] Pin numbers are not working properly when clicking near
            the center button

Change-Id: I412d4dcc6455e39703376c61e75d0150dd06dc3b
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/429123
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>