[n_upsable][browser][browsershell] Fix the zoom value is incorrect
:Release Notes:
The zoom value is not adaptive to the page view zoom factor

:Detailed Notes:
The zoom value is not adaptive to the page view zoom factor since the
'zoom_value' event is responded before the Zoom dropdown list has been
rendered. That makes the event 'zoom_value' listener does not execute to
update the current zoom value

:Testing Performed:
1. Launch the Enact Browser app
2. Visit google.com site
3. Click Zoom button and select 250
4. Toggle Zoom button multiple times
Observe that the active zoom value is always 250.

:QA Notes:
Test cases: QWTC-13359

:Issues Addressed:
[WRQ-33399] The zoom value is not adaptive to the page view zoom factor

Change-Id: I5408df9cf1570f72f36368e10a943b424c027dab
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/429263
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>