[n_upsable][browser][browsershell] Fix the bookmark dialog shows blank content
:Release Notes:
Sometimes the bookmark dialog shows blank content

:Detailed Notes:
Sometimes the bookmark dialog shows blank content when clicking the
add/remove bookmark icon. The root cause is that the showBookmarkDialog
is responded to before switchContent is responded to. Therefore, the
dialog does not receives the properties to display them.
Add 'switchedContent' event to ensure that switchContent is responded
to before calling showBookmarkDialog

:Testing Performed:
1. Launch the Enact Browser app
2. Visit https://w3schools.com site
3. Click Star icon to bookmark it
4. Click the Zoom icon
5. Multi-click toggle between 2 icon buttons
Observe that the bookmark dialog displays the content properly

:QA Notes:
Test cases: QWTC-12236

:Issues Addressed:
[WRR-4028] [TAS Failed][OSE][QWTC-12236] Verify that top 5 bookmarks
from "bookmark manager" is shown in the bookmark bar

Change-Id: I3ed8ac7c615affe3881a5856831f0614d6e71ad7
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/434979
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>