[n_upsable][browser][browsershell] Suggestion popup is not closed after touching outside of omnibox
:Release Notes:
Suggestion popup is not closed after touching outside of omnibox

:Detailed Notes:
The suggestion popup only closes after two times click because the
InputSpotlightDecorator component automatically locks pointer events
to manage focus with Spotlight. So two clicks are needed to trigger
the document click event.
To fix this, the noLockPointer property should be used to disable the
pointer lock behavior and need to customize Spotlight for the input.
Reference links:
http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/359990
http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/265729

:Testing Performed:
1. Launch the Enact Browser
2. Click on Address and input "Google"
Suggestion popup will display
4. Click outside of the omnibox
The suggestion popup is closed

:QA Notes:
Test-cases: QWTC-15056

:Issues Addressed:
[WRR-3584] [Touch Screen] Suggestion popup is not closed after
touching outside of omnibox

Change-Id: I3d341ffe7d425e2ead9a79085e29cda45241e0e1
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/448913
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>