[n_upsable][browser][browsershell] Add a valid check for shell.session.touch injection API
:Release Notes:
Add a check to verify the availability of the  shell.session.touch API,
injected from the web runtime, before attempting to use it.

:Detailed Notes:
The JavaScript code currently has logic to support and apply the
shell.session.touch API, which is injected from the web runtime.
However, when this API is not available from the web runtime, the
JavaScript code crashes.
This fix adds valid checks to ensure the API's availability before
attempting to use it.

:Testing Performed:
Local test successful

:QA Notes:
Test-cases: QWTC-15145

:Issues Addressed:
[WRR-3135] Implement and disable unavailable UIs/features

Change-Id: I5389094ac312bc584ed90e7bf972d7fc9c68fd7a
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/452896
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>