[n_upsable][browser][browsershell] Off site-filtering after resetting settings
:Release Notes:
The site-filtering is not disabled from the JS side after the user
resets the settings.

:Detailed Notes:
The site-filtering is not disabled from the JS side after the user
resets the settings. The webpages which are blocked in the previous step
are still prevented and cannot be loaded properly.
This patch fixes the error.

:Testing Performed:
1. Launch Enact Browser
2. Open Settings page > Site Filtering page
3. Select the blocked mode then add google.com to the blocked list
4. Open google.com and obsever that it is prevented,
an error page is displayed.
5. Back to Settings page and click to Reset Settings
6. Reload google.com in the previous step
Observe that the site is loaded successfully

:QA Notes:
Test cases: QWTC-15325

:Issues Addressed:
[WRR-11037] Can't access blocked websites after resetting settings

Change-Id: Icb83769eec7d0b5d8a608945ac3b89e06488c4c8
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/456999
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>