[n_upsable][browser][browsershell] Update the target parameter when launch/relaunch Enact
:Release Notes:
Retrieve the target parameter when relaunching the Enact Browser
properly to ensure compatibility with both OSE and Apollo

:Detailed Notes:
There is a slight difference in launch parameters between Apollo & OSE.
Therfore, we need to check the target parameter available in both before
running.
Additionally, nativeLifeCycleInterfaceVersion version 1 is not
compatible on Apollo. Therefore, it is necessary to define verions 2
in appinfo.json.
Refer to
http://collab.lge.com/main/display/~jiwon1.park/Native+app+life+cycle+interface

:Testing Performed:
1. Launch Enact Browser application
2. Visit www.google.com
3. Press Window keyboard to visit Home Screen
4. Click Enact Browser icon to open the app again
Observe Enact Browser opens with Google site

:QA Notes:
Test cases: QWTC-13630
Depends-on: http://gpro.lge.com/c/webosose/chromium120/+/457911

:Issues Addressed:
[WRR-10650] Enact Browser is re-opened while running in the background

Change-Id: I3185bc8d1cb713dad56dd77cc8063bc761d246c2
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/458272
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>