[browser][browsershell] Fix the bookmark dialog does not disappear
:Release Notes:
Fix the bookmark dialog doesn't disappear when clicking outside the area

:Detailed Notes:
There is a side effect in submissions/@17.browsershell.12.
The bookmark dialog does not disappear when clicking outside the area.
The cause of the issue is that the click listener is removed
immediately after clicking the Star button. Therefore, the code to
close the popup is not executed.

:Testing Performed:
1. Launch Enact Browser application
2. Visited any site, e.g. https://youtube.com
3. Click to the Star icon to bookmark this site
4. Click to a blank space outside of the bookmark popup
Observe that the bookmark dialog disappear

:QA Notes:
Test cases: QWTC-11155

:Issues Addressed:
[WRR-12510] The bookmark dialog does not work properly after syncing

Change-Id: I59e515ffd5bd81ef047e6b5028409fd743786f68
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/458440
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: Jeongran Jang <jeongran.jang@lge.com>