[np_upable][browser][browsershell] Fix bookmark star icon showing incorrect status
:Release Notes:
Update the default list of bookmarked URLs

:Detailed Notes:
The status of the Star icon is incorrect because the bookmarked URL does
not match the navigation URL after the page loads with full prefixes.
This patch updates the corresponding bookmarked URL.

:Testing Performed:
1. Launch Enact Browser
2. Open one of the 4 bookmarked URLs
3. Wait for the navigation URL to be updated
Observe that the Star icon is dark instead of white.

:QA Notes:
Test cases: QWTC-13288

:Issues Addressed:
[WRR-12845] [Automation][OSE][QWTC-13288] verify default bookmark sites
            status

Change-Id: I34279bcd0820ed0bbd9b5d5afb0f1431daa17068
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/459481
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Phan Thanh <thanh2.phan@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>