[n_upsable][browser][browsershell] Fix page doesn't load when clicking on a url suggestion item
:Release Notes:
Fix page doesn't load when clicking on a url suggestion item

:Detailed Notes:
The Suggestion URL Overlay is removed before the
click_suggested_item event is sent to the Main UI.
The isEditing state is responsible for opening and closing URL
suggestions.
Updating isEditing in the onActivate and onDeactivate functions is
intended to handle the opening and closing of URL suggestions during
Spotlight navigation.
However, this causes an incorrect processing order when navigating with
the mouse, as isEditing is updated before other events are triggered.
Therefore, the logic for handling Spotlight navigation should be moved
to the onKeyDown function.

:Testing Performed:
1. Launch Enact Browser
2. Access google.com
3. Click Url search
4. Enter h
5. Tap on any one in URL suggestions
6. Google page is shown after click

:QA Notes:
Test cases: QWTC-13109

:Issues Addressed:
[WRR-12724] [Automation][OSE][QWTC-13109] Verify show list of 6 URL
suggestions while entering URL.

Change-Id: I8fb92171dfa5f2e3891ae4a92612408b422cf73b
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/459468
Reviewed-by: Automated Checker <automated_checker@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>