[n_upsable][browser][browsershell] Fix abnormal UIOverlay popup
:Release Notes:
Fix error sometimes UI doesn't show full height of content

:Detailed Notes:
The height of the wrap element was being incorrectly retrieved from
getBoundingClientRect after the UIOverlay component received the new
state and updated the UI with the content corresponding.
This patch will add a timer that waits for the UI to render successfully
before calculating the height of the popup content.
This is a workaround to fix the issue, and the root cause has been
reported in WRR-13254
Furthermore, the 'Bookmark has been deleted' popup is only displayed
for a very short time. This comes from the asynchronous between
displaying and waiting to hide it.

:Testing Performed:
1. Launch Enact Browser application
2. Open www.mictests.com
3. Click the button in the web content to take a mic permission
Observe that the UserPermission popup appears properly and normally

:QA Notes:
Test cases: QWTC-13617

:Issues Addressed:
[WRR-10793] [browser_shell] Abnormal user permission popup UI

Change-Id: I592ec349da057918b80be15f4cf32cd0945b2cfb
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/460022
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>