[n_upsable][browser][browsershell] Update Clear Browsing Data popup content
:Release Notes:
Update Clear Browsing Data popup content to be clearer

:Detailed Notes:
The current description in the pop-up when clicking the CLEAR BROWSING
DATA button on the settings page is ambiguous from the end user's
perspective. It requires a more detailed explanation regarding which
data will be cleared.
Additionally, adding the content detail for multiple languages.

:Testing Performed:
1. Launch Enact Browser application
2. Open Settings page
3. Click Clear Browsing Data button to open the confirmation popup
Observe that the popup has a detail content

:QA Notes:
Test cases: QWTC-11173

:Issues Addressed:
[WRR-12480] Clarification of "Clear Browsing Data" Pop-Up Content

Change-Id: Ia621a8f36c86fa6a0efa3ba72572f5f46d004e2b
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/460326
Reviewed-by: Alexander Lapshenkov <alexander.lapshenkov@lge.com>
Reviewed-by: Automated Checker <automated_checker@lge.com>
Reviewed-by: AI Reviewer <ai_reviewer@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Jeonghee27 Ahn <jeonghee27.ahn@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>