[n_upsable][browser][browsershell] Update Reset Settings popup content
:Release Notes:
Update Reset Settings popup content to be clearer

:Detailed Notes:
The current description in the pop-up when clicking the RESET SETTINGS
button on the settings page is ambiguous from the end user's
perspective.It requires more detailed information, specifically
an exact list of what will be reset or cleared.
Additionally, update the content detail for multiple languages.

:Testing Performed:
1. Launch Enact Browser application
2. Open Settings page
3. Click Reset Settings button to open the confirmation popup
Observe that the popup has a detail content

:QA Notes:
Test cases: QWTC-13246

:Issues Addressed:
[WRR-12481] Clarification of "Reset Settings" Pop-Up Content

Change-Id: I9d2b0bb95803c958eacb04416d94494da86ecb1d
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.app.enactbrowser/+/460346
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>